Update E.13 Ordnance Survey mapping
We are pleased to have contributed to the discussion some ten years ago but recognise that the subject area has moved on. We would not recommend people starting to relate to our ontology now, and we look forward to migrating to some more authoritative one in due course.

Also, eastings & northings are best not confused with latitudes & longitudes. I have put in the definitions from the OS ontology: they are metres from an origin, in a defined "flat & square" projection.

I'm glad you've ignored the grid square classes in our ontologies!