1.1 alignments: Add missing prefixes and fix most typos
This patch addresses issues discovered when attempting to load all of
the alignment files into a Turtle file with an RDF processing tool.

One issue was NOT addressed: osiuk-alignments.ttl contains this line:

```turtle
spatialukgeom:AbstractGeometry geo:Geometry .
```

This is an incomplete triple, and as a non-member of the development
group for this repository, I did not know what aligning predicate would
be appropriate.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>