Incorporating changes to ConcaveHull description in specification
We're only handling the case of a ConcaveHull without any
parameters because various implementers use different parameters
and trying to handle all permutations of these is outside of the
scope of the standard. As such, implementers will elect to set
default values, and we wish to ensure that these are documented
by them for their users.

Fixes #270